requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
pydantic
pytest
httpx
python-dotenv